trace-cruncher: Refactor the way libtracefs instances are handled
Currently all libtracefs instances created by trace-cruncher are
wrapped by a dedicated C structure and stored in a binary search
tree. The ordering in the tree is based on a string comparison. The
tree is used for accessing the created instances and for cleanup
when the module exits (garbage collection). In this refactoring we
eliminate the need of the tree by switching to using the custom
Python type for libtracefs instances. The new type is defined in
the C extension of the module.

Link: https://lore.kernel.org/linux-trace-devel/20210819130827.12327-6-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>