trace-cruncher: Refactor the way kprobes are handled
Currently all kprobes, created by trace-cruncher, are stored in a
binary search tree. The ordering in the tree is based on a string
comparison. The tree is used for bookkeeping and to cleanup all
created kprobes when the module exits (garbage collection). In
this refactoring we eliminate the need of the tree by switching to
using the custom Python type for kprobes. The new type is defined
in the C extension of the module.

Link: https://lore.kernel.org/linux-trace-devel/20210819130827.12327-8-y.karadz@gmail.com

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>