Moving tests under tracecruncher/
This moves the tests directory under tracecruncher/ as otherwise the
tests/ directory gets installed in the top level of the python version
namespace which is not where it's expected (or really wanted) to live.

I can't find a specific PEP that defines this (probably more due to
ignorance than it not existing) but this ensures the tests/ directory
for trace-cruncher stays in the tracecruncher namespace and doesn't
polute outside of that.

Signed-off-by: John 'Warthog9' Hawley (VMware) <warthog9@eaglescrag.net>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>