trace-cruncher: Optimize get_arg_list()
The extraction of a 'C' string from a Python list of strings gets
defined as a static method instead of being coded in the body of
'get_arg_list()'. The new static method will be used in following
patches. The error message emitted in the case of an error is
remover since it will never get printed (it is overwritten by the
error messages of the functions calling 'get_arg_list()').

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>