trace-cruncher: Add static method for getting tep_handle object
The libtracefs APIs for getting the local (for the system) version of the
tep_handle object are being called in number of places in trace-cruncher.
In each of these places a hendling of the error case gets reimplemented,
resulting in a lot of duplicated code. Here we add a static method that
returns the object or prinst a standart error message end returns NULL
in a case of an error.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>