trace-cruncher: Rename python function argument in PyFtrace_eprobe
The keyword (named) argument of the eprobe constrictor is renamed from
'fetchargs' to 'fetch_fields'. This is done because the old name can be
confusing, since the probe is supposed to fetch a field of event (not a
function argument). The new name also makes a better match with the
'snake case' naming convention used in trace-cruncher.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>