trace-cruncher: Moving tests outside of tracecruncher/
This patch is a partial revert of commit:
086f784 (Moving tests under tracecruncher/)

The problem that the reverted commit tries to address is that tests have
been installed in the top level of the python version namespace which is
not where it's expected. Here we solve this problem by making sure the
tests are not installed at all. This is done by moving the 'tests'
directory back to the trunk of the repository and in the same time
removing the unnecessary '__init__.py' file from the top-level directory
of the tests (./tests/__init__.py). Note that we keep '__init__.py' files
in the sub-directories of 'tests'.

Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>