setup.py: fix string <-> int comparison
Slightly older (read: python 3.6ish) seem to treat the language_level
object as an integer all the time.  If you end up passing the version
in as a string (with the 's) it ends up doing a string to int comparison
and throws an error.  Newer python will deal with the non-matching
typing fine and this isn't an issue there.

So minor fix to be slightly more correct with relation to older python

Signed-off-by: John 'Warthog9' Hawley (VMware) <warthog9@eaglescrag.net>
Signed-off-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>