trace-cruncher: Fix CI issue and add reproducibility tools
This commit adds a scripts/ directory and the date-snapshot.sh script.
date-snapshot automates the process of checking out the last commit
in a repository before a specific date.

We leverage this tool in the GitHub actions CI to fix the ongoing build
error by locking our build dependencies to a specific date- currently
June 1, 2022. This date can be advanced when the breaking change is
found and fixed in the dependencies.

This commit is the beginning of work to make the trace-cruncher build
fully reproducible for containerization.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>