trace-cruncher: Implement PR feedback
This commit adds arguments to the date-snapshot script to enable deleting existing directories and either exiting or proceeding from there. This functionality is used in `make clean` to remove build artifacts. It also eases iterative testing when using the script.

In addition, the git source for kernel-shark has been changed to the mainline kernel.org repo because the old remote was used for testing and was not intended for stable release.

Finally, whitespace was cleaned up and invisible characters removed for easier readability.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>