trace-cruncher: Refactor and improve date-snapshot
This commit improves the behavior of date-snapshot in unexpected conditions and improves readability.

Explicit conditionals added as they are best practice to ensure consistent behavior.
In addition, quotes added around all inputs to the script to prevent multiple arguments from being passed to commands.
Finally, added a warning when a directory that is about to be deleted does not exist so that users are warned in unexpected situations.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>