trace-cruncher: Introduce Docker installation
This commit continues work towards making trace-cruncher fully reproducible and containerized.

The Dockerfile builds a Debian image with trace-cruncher installed, using the new date-snapshot script to freeze the built dependencies as well as trace-cruncher itself.

The next phase of this work will be moving to a Debuerreotype base image so that the container build will be completely reproducible.

Additionally this commit also updates the README with basic Docker install instructions.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>