trace-cruncher: Add more flags and workdir tracking to date-snapshot
This commit further extends the feature set of date-snapshot and helps
ensure it will behave as expected in all situations.

Added logging levels so that users can debug the behavior of the script.
By default, only critical errors are printed.

Added a --latest flag to bypass the specified date and get the latest version
of the repository. This will be leveraged in GitHub actions to check for
regressions or breaking changes automatically.

Added workdir tracking to the script. When the -f flag is passed, the workdir
is the parent dir of the provided file. When -i is passed, the workdir is the
current working directory.

Workdir tracking ensures commands like -d will always execute in the same
directory as the repos were created in, as long as the input file is the same.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>