trace-cruncher: Fix CI issue and add reproducibility tools (#3)
* trace-cruncher: Fix CI issue and add reproducibility tools

This commit adds a scripts/ directory and the date-snapshot.sh script.
date-snapshot automates the process of checking out the last commit
in a repository before a specific date.

We leverage this tool in the GitHub actions CI to fix the ongoing build
error by locking our build dependencies to a specific date- currently
June 1, 2022. This date can be advanced when the breaking change is
found and fixed in the dependencies.

This commit is the beginning of work to make the trace-cruncher build
fully reproducible for containerization.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>
Reviewed-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>

* trace-cruncher: Implement PR feedback

This commit adds arguments to the date-snapshot script to enable deleting existing directories and either exiting or proceeding from there. This functionality is used in `make clean` to remove build artifacts. It also eases iterative testing when using the script.

In addition, the git source for kernel-shark has been changed to the mainline kernel.org repo because the old remote was used for testing and was not intended for stable release.

Finally, whitespace was cleaned up and invisible characters removed for easier readability.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>
Reviewed-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>

* trace-cruncher: Refactor and improve date-snapshot

This commit improves the behavior of date-snapshot in unexpected conditions and improves readability.

Explicit conditionals added as they are best practice to ensure consistent behavior.
In addition, quotes added around all inputs to the script to prevent multiple arguments from being passed to commands.
Finally, added a warning when a directory that is about to be deleted does not exist so that users are warned in unexpected situations.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>
Reviewed-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>

* trace-cruncher: Add more flags and workdir tracking to date-snapshot

This commit further extends the feature set of date-snapshot and helps
ensure it will behave as expected in all situations.

Added logging levels so that users can debug the behavior of the script.
By default, only critical errors are printed.

Added a --latest flag to bypass the specified date and get the latest version
of the repository. This will be leveraged in GitHub actions to check for
regressions or breaking changes automatically.

Added workdir tracking to the script. When the -f flag is passed, the workdir
is the parent dir of the provided file. When -i is passed, the workdir is the
current working directory.

Workdir tracking ensures commands like -d will always execute in the same
directory as the repos were created in, as long as the input file is the same.

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>
Reviewed-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>