trace-cruncher: Finish integration of git-snapshot and fix issues
Fix an issue in Actions where the path of git-snapshot was not updated
properly

Fix a syntax error in git-snapshot causing the script to incorrectly interpret
the result of git rev-parse

Update the Dockerfile to account for the git-snapshot changes and switch
the Dockerfile to use the latest trace-cruncher while we are still
integrating Docker and finalizing the Dockerfile. Once we reach a stable
point this can be updated to use a snapshot again.

Update the Makefile with the correct path of git-snapshot

Signed-off-by: June Knauth (VMware) <june.knauth@gmail.com>