trace-cruncher: Add new ftraccepy API - wait()
Waiting for some conditions while the trace is running is a common
pattern in most of the trace-cruncher use cases. A new generic ftracepy
API is added to handle that functionality:
 ftracepy.wait()
It waits for any combination of these conditions:
 - a set of signals
 - until all pids of given list are allive
 - given amount of time.
This functionality is implemented in the trace-cruncher common library,
and not all of it is exposed through the python API. The underlying
internal API supports running of a custom callback, while waiting for
the given conditions. This will be used internally by trace-cruncher.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>