trace-cruncher: Fix waiting logic for external tasks
The ftracepy.wait() function has a logic to monitor given list of tasks.
The logic is different for children tasks and external processes. There
is a bug in monitoring external processes, the PID of a competed process
is incorrectly set to -1 instead of 0, as the type of the PID is
unsigned integer.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>