trace-cruncher: Enhance events filter API
Ftarce can filter events globally, per event system. The filters in
the event system filter file affect all events in that system. This is
very useful functionality, that's why the existing trace-cruncher event
filtering APIs are extended to support it. The 'event' parameter is
changed to optional. If event is missing, the APIs work with global
system filter file.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>