trace-cruncher: Replace permission error with warning
Access to the ftarce subsystem requires root privilege, trace-cruncher
cannot be used without it. There is a check for that at module load and
fatal error is triggered. However, there are cases where the module is
loaded for some other purposes as non-root, for example when the
documentation is generated. This should be possible without root
privilege.
Replaced the error with warning, so the module can be load by regular
user. However, a warning is printed in such cases.

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>