trace-cruncher: Improve trace instance deletion
Deletion of a trace instance is not always straightforward. If there is
a complex configuration, deletion may fail if the instance is not in its
default state. Recently a new API was introduce in the tracefs library,
which resets a trace instance to its default state. That new API is used
to improve the logic of trace instance deletion. A new set of
trace-cruncher APIs is introduced, methods of the instance object:
 reset(), delete()

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>